+++ src/main/java/gg/vape/mapping/EventPreTickCallback.java
@@ -9,4 +9,3 @@ extends InsertedTickCallbackBase {
         new EventPreTick().fire();
     }
 }
-